# bushop-scraper/openapi.json
{"openapi":"3.1.0","info":{"title":"BusHop Scraper API","version":"0.1.0"},"paths":{"/api/v1/routes":{"get":{"tags":["routes"],"summary":"Get Routes","description":"Returns all routes with i18n province names. Phase 1: empty list.","operationId":"get_routes_api_v1_routes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse_list_RouteResponse__"}}}}}}},"/api/v1/schedules":{"get":{"tags":["schedules"],"summary":"Get Schedules","description":"Search bus schedules. Phase 1: empty list.","operationId":"get_schedules_api_v1_schedules_get","parameters":[{"name":"origin_id","in":"query","required":true,"schema":{"type":"integer","description":"ID of departure province","title":"Origin Id"},"description":"ID of departure province"},{"name":"destination_id","in":"query","required":true,"schema":{"type":"integer","description":"ID of destination province","title":"Destination Id"},"description":"ID of destination province"},{"name":"date","in":"query","required":true,"schema":{"type":"string","description":"Travel date in YYYY-MM-DD format","title":"Date"},"description":"Travel date in YYYY-MM-DD format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse_list_ScheduleResponse__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scrape-jobs":{"post":{"tags":["scrape-jobs"],"summary":"Create Scrape Job","description":"Create a scrape job. Phase 1: placeholder.","operationId":"create_scrape_job_api_v1_scrape_jobs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeJobCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse_ScrapeJobResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scrape-jobs/{job_id}":{"get":{"tags":["scrape-jobs"],"summary":"Get Scrape Job","description":"Get scrape job status. Phase 1: placeholder.","operationId":"get_scrape_job_api_v1_scrape_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse_ScrapeJobResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DataResponse_ScrapeJobResponse_":{"properties":{"data":{"$ref":"#/components/schemas/ScrapeJobResponse"},"meta":{"$ref":"#/components/schemas/MetaResponse","default":{}}},"type":"object","required":["data"],"title":"DataResponse[ScrapeJobResponse]"},"DataResponse_list_RouteResponse__":{"properties":{"data":{"items":{"$ref":"#/components/schemas/RouteResponse"},"type":"array","title":"Data"},"meta":{"$ref":"#/components/schemas/MetaResponse","default":{}}},"type":"object","required":["data"],"title":"DataResponse[list[RouteResponse]]"},"DataResponse_list_ScheduleResponse__":{"properties":{"data":{"items":{"$ref":"#/components/schemas/ScheduleResponse"},"type":"array","title":"Data"},"meta":{"$ref":"#/components/schemas/MetaResponse","default":{}}},"type":"object","required":["data"],"title":"DataResponse[list[ScheduleResponse]]"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MetaResponse":{"properties":{"scraped_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scraped At"}},"type":"object","title":"MetaResponse"},"ProvinceI18n":{"properties":{"vi":{"type":"string","title":"Vi"},"en":{"type":"string","title":"En"},"ja":{"type":"string","title":"Ja"}},"type":"object","required":["vi","en","ja"],"title":"ProvinceI18n"},"RouteResponse":{"properties":{"id":{"type":"integer","title":"Id"},"slug":{"type":"string","title":"Slug"},"origin_id":{"type":"integer","title":"Origin Id"},"destination_id":{"type":"integer","title":"Destination Id"},"origin_name":{"$ref":"#/components/schemas/ProvinceI18n"},"destination_name":{"$ref":"#/components/schemas/ProvinceI18n"}},"type":"object","required":["id","slug","origin_id","destination_id","origin_name","destination_name"],"title":"RouteResponse"},"ScheduleResponse":{"properties":{"id":{"type":"integer","title":"Id"},"route_id":{"type":"integer","title":"Route Id"},"operator_name":{"type":"string","title":"Operator Name"},"departure_time":{"type":"string","title":"Departure Time"},"arrival_time":{"type":"string","title":"Arrival Time"},"price":{"type":"integer","title":"Price"},"scraped_at":{"type":"string","title":"Scraped At"},"is_stale":{"type":"boolean","title":"Is Stale"}},"type":"object","required":["id","route_id","operator_name","departure_time","arrival_time","price","scraped_at","is_stale"],"title":"ScheduleResponse"},"ScrapeJobCreate":{"properties":{"route_id":{"type":"integer","title":"Route Id"}},"type":"object","required":["route_id"],"title":"ScrapeJobCreate"},"ScrapeJobResponse":{"properties":{"id":{"type":"integer","title":"Id"},"status":{"type":"string","enum":["pending","running","completed","failed"],"title":"Status"},"route_id":{"type":"integer","title":"Route Id"},"created_at":{"type":"string","title":"Created At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","status","route_id","created_at"],"title":"ScrapeJobResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}